Rename NavBar component to NavBar for easier debugging
